MSVC: Add project files for MSVC2019
Microsoft removed (or at least deprecated) support for global per-user
property sheets. This was a nice feature to tell MSVC where to look for
includes/libraries that are not part of the project (like SDL, png,
etc.)

Removal of this feature means that all property sheets need to be part
of the project/solution. Since I don't want to force everyone to use the
same include/library paths than I do, I had to use a workaround.

I added a property sheet "localdirs.props" that adds the include and
library paths from the environment variables INCLUDE and LIB32
respectively (64bit build is still broken currently, so only LIB32 is
needed)

Please set up these variables before opening/building the solution, so
that your local libs and include files can be found.

If anyone finds a smarter solution to this, feel free to submit a PR :)